Remove bundled pathlib and the scandir dependency for PY2
